new changes

+++ ft_memcmp.c
@@ -6,7 +6,7 @@
 /*   By: hramirez <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/02/21 14:54:53 by hramirez          #+#    #+#             */
-/*   Updated: 2018/03/01 18:49:12 by hramirez         ###   ########.fr       */
+/*   Updated: 2018/03/13 17:20:43 by hramirez         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -16,7 +16,7 @@ int					ft_memcmp(const void *s1, const void *s2, size_t n)
 {
 	unsigned char	*str_1copy;
 	unsigned char	*str_2copy;
-	int				index;
+	size_t			index;
 
 	str_1copy = (unsigned char *)s1;
 	str_2copy = (unsigned char *)s2;
