+++ ft_memmove.c
@@ -6,7 +6,7 @@
 /*   By: hramirez <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/02/20 16:47:39 by hramirez          #+#    #+#             */
-/*   Updated: 2018/02/26 14:08:01 by hramirez         ###   ########.fr       */
+/*   Updated: 2018/03/13 17:19:54 by hramirez         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -16,7 +16,7 @@ void				*ft_memmove(void *dst, const void *src, size_t n)
 {
 	unsigned char	*dst_copy;
 	unsigned char	*src_copy;
-	int				index;
+	size_t			index;
 
 	dst_copy = (unsigned char *)dst;
 	src_copy = (unsigned char *)src;
